修改配方中时间相关参数为小数，兼容需要调节0.1单位。

+++ common/src/main/java/com/inno/common/db/entity/FormulaEntity.kt
@@ -76,7 +76,7 @@ data class Formula(
 sealed class FormulaItem {
     @Serializable
     data class FormulaUnitValue(
-        var value: Short,
+        var value: Float,
         var rangeStart: Float,
         var rangeEnd: Float,
         var unit: String,
